Continue update process with Tracker2.1.3.  references #1265

+++ fusioninventory/tools/phpunit/TestInstallUpdate.php
@@ -1,5 +1,10 @@
 <?php
 
+/*
+ *  You can run this file with arg dbread to use current DB and not write DB tests like
+ * phpunit TestInstallUpdate.php dbread
+ */
+
 define('PHPUnit_MAIN_METHOD', 'Plugins_Fusioninventory_TestInstallUpdate::main');
 
 if (!defined('GLPI_ROOT')) {
@@ -159,27 +164,31 @@ class Plugins_Fusioninventory_TestInstallUpdate extends PHPUnit_Framework_TestCa
    function testDB() {
       global $DB;
        
-      $query = "SHOW TABLES";
-      $result = $DB->query($query);
-      while ($data=$DB->fetch_array($result)) {
-         if (strstr($data[0], "tracker")
-                 OR strstr($data[0], "fusi")) {
-            $DB->query("DROP TABLE ".$data[0]);
+      if (isset($argv[1]) 
+              AND $argv[1] == 'dbread') {
+         // Not write DB, so use current DB 
+      } else {
+         $query = "SHOW TABLES";
+         $result = $DB->query($query);
+         while ($data=$DB->fetch_array($result)) {
+            if (strstr($data[0], "tracker")
+                    OR strstr($data[0], "fusi")) {
+               $DB->query("DROP TABLE ".$data[0]);
+            }
+         }
+      
+         // ** Insert in DB
+         $version = "2.3.3";
+         $version = "2.1.3";
+         $DB_file = GLPI_ROOT ."/plugins/fusioninventory/tools/phpunit/dbupdate/i-".$version.".sql";
+         $DBf_handle = fopen($DB_file, "rt");
+         $sql_query = fread($DBf_handle, filesize($DB_file));
+         fclose($DBf_handle);
+         foreach ( explode(";\n", "$sql_query") as $sql_line) {
+            if (get_magic_quotes_runtime()) $sql_line=stripslashes_deep($sql_line);
+            if (!empty($sql_line)) $DB->query($sql_line);
          }
       }
-      
-      
-      // ** Insert in DB
-      $version = "2.3.3";
-      $version = "2.1.3";
-      $DB_file = GLPI_ROOT ."/plugins/fusioninventory/tools/phpunit/dbupdate/i-".$version.".sql";
-      $DBf_handle = fopen($DB_file, "rt");
-      $sql_query = fread($DBf_handle, filesize($DB_file));
-      fclose($DBf_handle);
-      foreach ( explode(";\n", "$sql_query") as $sql_line) {
-         if (get_magic_quotes_runtime()) $sql_line=stripslashes_deep($sql_line);
-         if (!empty($sql_line)) $DB->query($sql_line);
-      }       
        
       passthru("cd .. && /usr/local/bin/php -f cli_install.php");
